requirements.txt
astroid==2.1.0
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.2
Django==2.1.7
djangorestframework==3.9.1
djangorestframework-csv==2.1.0
idna==2.8
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
mccabe==0.6.1
pylint==2.2.2
pylint-django==2.0.5
pylint-plugin-utils==0.4
pytz==2018.9
requests==2.21.0
six==1.12.0
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.24.1
wrapt==1.11.1
